+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/HDGlobalSettingsWindow.cs
@@ -84,11 +84,16 @@ namespace UnityEditor.Rendering.HighDefinition
 
             if (serializedSettings == null || settingsSerialized != HDRenderPipelineGlobalSettings.instance)
             {
-                if (HDRenderPipeline.currentAsset != null || HDRenderPipelineGlobalSettings.instance != null)
+                if (HDRenderPipelineGlobalSettings.instance != null)
                 {
                     settingsSerialized = HDRenderPipelineGlobalSettings.Ensure();
                     var serializedObject = new SerializedObject(settingsSerialized);
                     serializedSettings = new SerializedHDRenderPipelineGlobalSettings(serializedObject);
+                }
+                else
+                {
+                    serializedSettings = null;
+                    settingsSerialized = null;
                 }
             }
             else if (settingsSerialized != null && serializedSettings != null)
@@ -123,13 +128,27 @@ namespace UnityEditor.Rendering.HighDefinition
 
             if (isHDRPinUse)
             {
-                EditorGUILayout.HelpBox(Styles.warningGlobalSettingsMissing, MessageType.Warning);
+                ShowMessageWithFixButton(Styles.warningGlobalSettingsMissing, MessageType.Warning);
             }
             else
             {
                 EditorGUILayout.HelpBox(Styles.warningHdrpNotActive, MessageType.Warning);
                 if (serialized == null)
-                    EditorGUILayout.HelpBox(Styles.infoGlobalSettingsMissing, MessageType.Info);
+                {
+                    ShowMessageWithFixButton(Styles.infoGlobalSettingsMissing, MessageType.Info);
+                }
+            }
+        }
+
+        void ShowMessageWithFixButton(string helpBoxLabel, MessageType type)
+        {
+            using (new EditorGUILayout.HorizontalScope())
+            {
+                EditorGUILayout.HelpBox(helpBoxLabel, type);
+                if (GUILayout.Button(Styles.fixAssetButtonLabel, GUILayout.Width(45)))
+                {
+                    HDRenderPipelineGlobalSettings.Ensure();
+                }
             }
         }
 
@@ -145,6 +164,7 @@ namespace UnityEditor.Rendering.HighDefinition
                 if (EditorGUI.EndChangeCheck())
                 {
                     HDRenderPipelineGlobalSettings.UpdateGraphicsSettings(newAsset);
+                    Debug.Assert(newAsset == HDRenderPipelineGlobalSettings.instance);
                     if (settingsSerialized != null && !settingsSerialized.Equals(null))
                         EditorUtility.SetDirty(settingsSerialized);
                 }